requirements.txt
notion-client>=2.0.0
anthropic>=0.13.0
claude-code-sdk
python-dotenv>=1.0.0
requests>=2.31.0
nest-asyncio>=1.5.8
aiohttp>=3.9.1
gitpython>=3.1.0
pydantic>=2.8.0
anyio 